+++ Controller.php
@@ -29,15 +29,7 @@ class Controller extends ControllerAdmin
         $view = new View('@ReferrersManager/index');
         $this->setBasicVariablesView($view);
 
-        $view->searchEngineInfos = Model::getInstance()->getSearchEngineInfos();
-        $view->searchEngineLogos = Model::getInstance()->getSearchEngineLogos();
-
-        $view->socialInfos = Model::getInstance()->getSocialsInfos();
-        $view->socialLogos = Model::getInstance()->getSocialsLogos();
-
         $view->ownSocialDefinitions = Model::getInstance()->areDefaultSocialsDisabled();
-        $view->userDefinedSocials = Model::getInstance()->getUserDefinedSocials();
-        $view->userDefinedSearchEngines = Model::getInstance()->getUserDefinedSearchEngines();
 
         return $view->render();
     }
@@ -47,6 +39,8 @@ class Controller extends ControllerAdmin
      */
     public function setDefaultSocialsDisabled()
     {
+        Piwik::checkUserHasSuperUserAccess();
+
         $state = Common::getRequestVar('state', 0, 'int');
         Model::getInstance()->setDefaultSocialsDisabled((bool) $state);
         Json::sendHeaderJSON();
@@ -60,7 +54,7 @@ class Controller extends ControllerAdmin
      */
     public function checkUrl()
     {
-        $this->checkTokenInUrl();
+        Piwik::checkUserHasSuperUserAccess();
 
         $urlToCheck = trim(Common::unsanitizeInputValue(Common::getRequestVar('url', null, 'string')));
 
@@ -73,7 +67,7 @@ class Controller extends ControllerAdmin
 
     public function refresh()
     {
-        $this->checkTokenInUrl();
+        Piwik::checkUserHasSuperUserAccess();
 
         Json::sendHeaderJSON();
 
@@ -85,103 +79,6 @@ class Controller extends ControllerAdmin
             Model::getInstance()->clearSearchEngineCache();
         }
 
-        return 1;
-    }
-
-    /**
-     * Ajax Action to add a news social network
-     * @return int
-     */
-    public function addSocial()
-    {
-        $this->checkTokenInUrl();
-
-        Json::sendHeaderJSON();
-
-        $name = Common::getRequestVar('name', '', 'string');
-        $host = Common::getRequestVar('host', '', 'string');
-
-        if (empty($host) || empty($name)) {
-            return 0;
-        }
-
-        $socials = Model::getInstance()->getUserDefinedSocials();
-        $socials[$host] = $name;
-        Model::getInstance()->setUserDefinedSocials($socials);
-        return 1;
-    }
-
-    /**
-     * Ajax action to remove a user defined social
-     * @return int
-     */
-    public function removeSocial()
-    {
-        $this->checkTokenInUrl();
-
-        Json::sendHeaderJSON();
-
-        $host = Common::getRequestVar('host', '', 'string');
-
-        if (empty($host)) {
-            return 0;
-        }
-
-        $socials = Model::getInstance()->getUserDefinedSocials();
-        unset($socials[$host]);
-        Model::getInstance()->setUserDefinedSocials($socials);
-        return 1;
-    }
-
-    /**
-     * Ajax action to add a ne search engine definition
-     * @return int
-     */
-    public function addSearchEngine()
-    {
-        $this->checkTokenInUrl();
-
-        Json::sendHeaderJSON();
-
-        $name = Common::getRequestVar('name', '', 'string');
-        $host = Common::getRequestVar('host', '', 'string');
-        $backlink = Common::getRequestVar('backlink', '', 'string');
-        $parameters = Common::getRequestVar('parameters', '', 'string');
-        $charset = Common::getRequestVar('charset', '', 'string');
-
-        if (empty($host) || empty($name)) {
-            return 0;
-        }
-
-        if (!empty($parameters)) {
-            $parameters = explode(',', $parameters);
-        }
-
-        $engines = Model::getInstance()->getUserDefinedSearchEngines();
-        $engines[$host] = array($name, $parameters, $backlink, $charset);
-        Model::getInstance()->setUserDefinedSearchEngines($engines);
-        return 1;
-    }
-
-    /**
-     * Ajax action to remove a user defined search engine definition
-     * @return int
-     */
-    public function removeSearchEngine()
-    {
-        $this->checkTokenInUrl();
-
-        Json::sendHeaderJSON();
-
-        $host = Common::getRequestVar('host', '', 'string');
-
-        if (empty($host)) {
-            return 0;
-        }
-
-        $engines = Model::getInstance()->getUserDefinedSearchEngines();
-        unset($engines[$host]);
-        Model::getInstance()->setUserDefinedSearchEngines($engines);
         return 1;
     }
 }