+++ CHANGELOG.md
@@ -1,33 +1,50 @@
 ## Changelog
 
+__3.0__
+
+* Compatibility for Piwik 3.x
+* Rewrote UI using angular JS
+* Materialised UI/UX
+* Searchable lists for search engines / social networks
+
 __1.9__
+
 * Mark plugin as incompatible with 3.x
 * Translation updates
 
 __1.8__
+
 * fixes minor bugs
 * adds possibility to refresh lists (clears caches)
 
 __1.7__
+
 * Fix Compatibility with Piwik 2.16
 
 __1.6__
+
 * Compatibility for Piwik > 2.15 / translation updates
 
 __1.5__
+
 * Compatibility issues for older Piwik versions
 
 __1.4__
+
 * Translation updates
 
 __1.3__
+
 * Adjustments for new Piwik menu api
 
 __1.2__
+
 * Compatibility fix for PHP < 5.4
 
 __1.1__
+
 * Minor Changes
 
 __1.0__
+
 * Initial release